LayoutTests/system-preview/model/model-element-expected.txt


PASS HTMLModelElement is exposed on Window.
PASS document.createElement('model') returns an HTMLModelElement object.
